Moving devshell module to contrib.
See discussion in #370 for context.
