+++ docs/source/oauth2client.contrib.devshell.rst
@@ -1,0 +1,7 @@
+oauth2client.contrib.devshell module
+====================================
+
+.. automodule:: oauth2client.contrib.devshell
+    :members:
+    :undoc-members:
+    :show-inheritance:
