+++ docs/source/oauth2client.contrib.rst
@@ -14,6 +14,7 @@ Submodules
 .. toctree::
 
    oauth2client.contrib.appengine
+   oauth2client.contrib.devshell
    oauth2client.contrib.django_orm
    oauth2client.contrib.flask_util
    oauth2client.contrib.gce
