--- docs/source/oauth2client.devshell.rst
@@ -1,7 +1,0 @@
-oauth2client.devshell module
-============================
-
-.. automodule:: oauth2client.devshell
-    :members:
-    :undoc-members:
-    :show-inheritance:
